chore: refactor profile user page to minimize the amount of code there is for the page it self
